split test files
add small delay
